Split up enum queries so they can be separately cached.
