+++ src-electron/db/query-zcl.js
@@ -23,95 +23,7 @@
 const dbApi = require('./db-api')
 const dbMapping = require('./db-mapping')
 const queryAtomic = require('./query-atomic')
-
-/**
- * Retrieves all the enums in the database.
- *
- * @export
- * @param {*} db
- * @returns Promise that resolves with the rows of enums.
- */
-async function selectAllEnums(db, packageId) {
-  return dbApi
-    .dbAll(
-      db,
-      `SELECT ENUM_ID, NAME, TYPE FROM ENUM  WHERE PACKAGE_REF = ? ORDER BY NAME`,
-      [packageId]
-    )
-    .then((rows) => rows.map(dbMapping.map.enum))
-}
-
-/**
- * Retrieves all the enums in the database.
- *
- * @export
- * @param {*} db
- * @returns Promise that resolves with the rows of enums.
- */
-async function selectClusterEnums(db, packageId, clusterId) {
-  return dbApi
-    .dbAll(
-      db,
-      `
-SELECT
-  E.ENUM_ID,
-  E.NAME,
-  E.TYPE
-FROM
-  ENUM AS E
-INNER JOIN
-  ENUM_CLUSTER AS EC
-ON
-  E.ENUM_ID = EC.ENUM_REF
-WHERE
-  E.PACKAGE_REF = ?
-  AND EC.CLUSTER_REF = ?
-ORDER BY E.NAME`,
-      [packageId, clusterId]
-    )
-    .then((rows) => rows.map(dbMapping.map.enum))
-}
-
-async function selectAllEnumItemsById(db, id) {
-  return dbApi
-    .dbAll(
-      db,
-      'SELECT NAME, VALUE FROM ENUM_ITEM WHERE ENUM_REF = ? ORDER BY FIELD_IDENTIFIER',
-      [id]
-    )
-    .then((rows) => rows.map(dbMapping.map.enumItem))
-}
-
-async function selectAllEnumItems(db, packageId) {
-  return dbApi
-    .dbAll(
-      db,
-      `SELECT ENUM_ITEM.NAME,
-              ENUM_ITEM.VALUE,
-              ENUM_ITEM.ENUM_REF
-       FROM ENUM_ITEM, ENUM
-       WHERE ENUM.PACKAGE_REF = ? AND ENUM.ENUM_ID = ENUM_ITEM.ENUM_REF
-       ORDER BY ENUM_ITEM.ENUM_REF, ENUM_ITEM.FIELD_IDENTIFIER`,
-      [packageId]
-    )
-    .then((rows) => rows.map(dbMapping.map.enumItem))
-}
-
-async function selectEnumById(db, id) {
-  return dbApi
-    .dbGet(db, 'SELECT ENUM_ID, NAME, TYPE FROM ENUM WHERE ENUM_ID = ?', [id])
-    .then(dbMapping.map.enum)
-}
-
-async function selectEnumByName(db, name, packageId) {
-  return dbApi
-    .dbGet(
-      db,
-      'SELECT ENUM_ID, NAME, TYPE FROM ENUM WHERE NAME = ? AND PACKAGE_REF = ? ORDER BY NAME',
-      [name, packageId]
-    )
-    .then(dbMapping.map.enum)
-}
+const queryEnum = require('./query-enum')
 
 /**
  * Retrieves all the bitmaps in the database.
@@ -1363,13 +1275,6 @@ async function updateDeviceTypeEntityReferences(db, packageId) {
 }
 
 // exports
-exports.selectAllEnums = selectAllEnums
-exports.selectClusterEnums = selectClusterEnums
-exports.selectAllEnumItemsById = selectAllEnumItemsById
-exports.selectAllEnumItems = selectAllEnumItems
-exports.selectEnumById = selectEnumById
-exports.selectEnumByName = selectEnumByName
-
 exports.selectAllBitmaps = selectAllBitmaps
 exports.selectClusterBitmaps = selectClusterBitmaps
 exports.selectAllBitmapFields = selectAllBitmapFields
@@ -1436,3 +1341,10 @@ exports.selectAllAtomics = queryAtomic.selectAllAtomics
 exports.selectAtomicSizeFromType = queryAtomic.selectAtomicSizeFromType
 exports.selectAtomicType = queryAtomic.selectAtomicType
 exports.selectAtomicById = queryAtomic.selectAtomicById
+
+exports.selectAllEnums = queryEnum.selectAllEnums
+exports.selectClusterEnums = queryEnum.selectClusterEnums
+exports.selectAllEnumItemsById = queryEnum.selectAllEnumItemsById
+exports.selectAllEnumItems = queryEnum.selectAllEnumItems
+exports.selectEnumById = queryEnum.selectEnumById
+exports.selectEnumByName = queryEnum.selectEnumByName
